# docs/api.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CAN-HMI Signal API",
    "description": "Real-time CAN bus signal monitoring and control API",
    "version": "0.1.0"
  },
  "paths": {
    "/signals": {
      "get": {
        "tags": [
          "Signals"
        ],
        "summary": "List latest signal values",
        "operationId": "list_signals_signals_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignalListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
...
    "/signals/{signal_name}": {
      "get": {
        "tags": [
          "Signals"
        ],
        "summary": "Get latest value for one signal",
        "operationId": "get_signal_signals__signal_name__get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "signal_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Signal Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignalValueResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Signals"
        ],
        "summary": "Write value to signal (CAN write)",
        "operationId": "write_signal_signals__signal_name__put",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "signal_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Signal Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WriteSignalRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/devmode/catalog": {
      "get": {
        "tags": [
          "Dev Mode"
        ],
        "summary": "Dev Mode signal families and selectable states",
        "description": "Metadata the frontend uses to build one tab and its state buttons per signal family.",
        "operationId": "devmode_catalog_api_devmode_catalog_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/devmode/status": {
      "get": {
        "tags": [
          "Dev Mode"
        ],
        "summary": "Current Dev Mode seat locks",
        "operationId": "devmode_status_api_devmode_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/devmode/seats/select": {
      "post": {
        "tags": [
          "Dev Mode"
        ],
        "summary": "Select seats for Dev Mode (locks other sections out)",
        "operationId": "select_seats_api_devmode_seats_select_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DevModeSeatSelectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/devmode/exit": {
      "post": {
        "tags": [
          "Dev Mode"
        ],
        "summary": "Leave Dev Mode and release all seat locks of this section",
        "operationId": "exit_devmode_api_devmode_exit_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/devmode/signals": {
      "post": {
        "tags": [
          "Dev Mode"
        ],
        "summary": "Apply one signal family to several seats at once",
        "operationId": "apply_devmode_signal_api_devmode_signals_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DevModeSignalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccessWarning": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Warning or access error code"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Short description for the frontend"
          },
          "profile_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Name",
            "description": "Profile name currently in effect"
          },
          "required_permission": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Required Permission",
            "description": "Permission required for the operation"
          },
          "signal_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signal Name",
            "description": "Affected signal name if this is a single-signal warning"
          },
          "signals": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Signals",
            "description": "List of signals skipped in the batch"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "AccessWarning",
        "description": "Access/profile warning information for a profile-scoped operation."
      },
...
      "DevModeSeatSelectRequest": {
        "properties": {
          "seats": {
            "additionalProperties": {
              "type": "boolean"
            },
            "type": "object",
            "title": "Seats",
            "description": "Map seat_id → selected (fl, fr, rl1, rl2, rr1)"
          },
          "block_timeout_sec": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 3600.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Block Timeout Sec",
            "description": "How long other sections stay blocked from writing the seat (seconds); defaults to 60"
          }
        },
        "type": "object",
        "required": [
          "seats"
        ],
        "title": "DevModeSeatSelectRequest",
        "description": "Select/deselect seats in Dev Mode — POST /api/devmode/seats/select."
      },
      "DevModeSignalRequest": {
        "properties": {
          "signal_name": {
            "type": "string",
            "title": "Signal Name",
            "description": "Signal family: ACR_RetractRequest | ABL_RetractRequest | ISB_Color | HB_Request"
          },
          "value": {
            "type": "number",
            "title": "Value",
            "description": "Value applied to every selected seat"
          },
          "seats": {
            "additionalProperties": {
              "type": "boolean"
            },
            "type": "object",
            "title": "Seats",
            "description": "Map seat_id → whether the value is applied"
          },
          "block_timeout_sec": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 3600.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Block Timeout Sec",
            "description": "Seat lock renewal duration (seconds); defaults to 60"
          }
        },
        "type": "object",
        "required": [
          "signal_name",
          "value",
          "seats"
        ],
        "title": "DevModeSignalRequest",
        "description": "Apply one signal family to several seats at once — POST /api/devmode/signals."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SignalListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SignalValueResponse"
            },
            "type": "array",
            "title": "Items",
            "description": "List of signals"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total number of signals in the list"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/AccessWarning"
            },
            "type": "array",
            "title": "Warnings",
            "description": "Access warnings, if any"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "SignalListResponse",
        "description": "List of signal values returned by the API."
      },
...
      "SignalValueResponse": {
        "properties": {
          "signal_name": {
            "type": "string",
            "title": "Signal Name",
            "description": "Unique signal identifier"
          },
          "std_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Std Name",
            "description": "Standard signal name; currently identical to signal_name"
          },
          "value": {
            "type": "number",
            "title": "Value",
            "description": "Decoded real value"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit",
            "description": "Measurement unit (for example: km/h, °C)"
          },
          "timestamp": {
            "type": "number",
            "title": "Timestamp",
            "description": "Unix timestamp (seconds) when the value was read"
          }
        },
        "type": "object",
        "required": [
          "signal_name",
          "value",
          "timestamp"
        ],
        "title": "SignalValueResponse",
        "description": "Current value of a CAN signal."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WriteSignalRequest": {
        "properties": {
          "value": {
            "type": "number",
            "title": "Value",
            "description": "Value to write to the CAN bus"
          }
        },
        "type": "object",
        "required": [
          "value"
        ],
        "title": "WriteSignalRequest",
        "description": "Request to write a value onto the CAN bus."
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}